apps/core/requirements.txt
# ORION Core Dependencies
# Python 3.10+

# DevOps Agent (for Action subsystem)
# Install DevOps Agent first: cd ../devops-agent && pip install -e .
# OR use: pip install -e ../devops-agent

# Web Framework
fastapi>=0.104.0
# Using plain uvicorn without [standard] to avoid uvloop issues in Docker
uvicorn>=0.24.0
websockets>=12.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Common utilities (shared with other ORION components)
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Integration
openai>=1.3.0           # For OpenAI-compatible API (vLLM)

# Async utilities
asyncio>=3.4.3
aiofiles>=23.2.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging
structlog>=23.2.0

# SSH & Docker control
paramiko>=3.3.0
docker>=6.1.0

# Data handling
python-multipart>=0.0.6
jinja2>=3.1.2

# Monitoring
psutil>=5.9.0
prometheus-client>=0.19.0

# Tracing (AI Toolkit integration)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-http>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-aiohttp-client>=0.42b0

# Telegram Bot
python-telegram-bot>=20.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
ruff>=0.1.0
